requirements.txt
colorama==0.4.6
kafka-python==2.0.2
loguru==0.7.3
python-dotenv==1.0.1
setuptools==75.8.0
six==1.17.0
wheel==0.45.1
win32_setctime==1.2.0
